move Icon component into its own panel util common file
