+++ packages/editor/src/components/panels/IconComponent.tsx
@@ -1,0 +1,43 @@
+/*
+CPAL-1.0 License
+
+The contents of this file are subject to the Common Public Attribution License
+Version 1.0. (the"License"); you may not use this file except in compliance
+with the License. You may obtain a copy of the License at
+https://github.com/ir-engine/ir-engine/blob/dev/LICENSE.
+The License is based on the Mozilla Public License Version 1.1, but Sections 14
+and 15 have been added to cover use of software over a computer network and 
+provide for limited attribution for the Original Developer. In addition, 
+Exhibit A has been modified to be consistent with Exhibit B.
+
+Software distributed under the License is distributed on an"AS IS" basis,
+WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for the
+specific language governing rights and limitations under the License.
+
+The Original Code is Infinite Reality Engine.
+
+The Original Developer is the Initial Developer. The Initial Developer of the
+Original Code is the Infinite Reality Engine team.
+
+All portions of the code written by the Infinite Reality Engine team are Copyright © 2021-2023 
+Infinite Reality Engine. All Rights Reserved.
+*/
+
+import { Entity, entityExists, getAllComponents } from '@ir-engine/ecs'
+import { getState } from '@ir-engine/hyperflux'
+import TransformPropertyGroup from '@ir-engine/ui/src/components/editor/properties/transform'
+import React from 'react'
+import { ComponentEditorsState } from '../../services/ComponentEditors'
+
+export function IconComponent({ entity }: { entity: Entity }) {
+  const icons = entityExists(entity)
+    ? getAllComponents(entity)
+        .map((c) => getState(ComponentEditorsState)[c.name]?.iconComponent)
+        .filter((icon) => !!icon)
+    : []
+  const _IconComponent = icons.length > 0 ? icons[0] : TransformPropertyGroup.iconComponent
+  if (!_IconComponent) return null
+  return (
+    <_IconComponent entity={entity} className="h-5 w-5 flex-shrink-0" data-testid="hierarchy-panel-scene-item-icon" />
+  )
+}
